Move ns column functions to utils file
