--- src/shared/components/namespaceColumns/createNamespaceColumnName.ts
@@ -1,8 +1,0 @@
-import _ from 'lodash';
-
-export function createNamespaceColumnName(
-    namespaceName: string,
-    namespaceColumnName: string
-) {
-    return namespaceName + ' ' + _.capitalize(namespaceColumnName);
-}
